fix(rust): write score as f64 float instead of i64 integer in pick/allocate-task
Rust was truncating the score to i64 before writing to YAML, producing
integer values (e.g. 123) instead of floats (e.g. 123.0) as Python does.
Zig was already correct — its {d:.1} format writes unquoted YAML floats.

Closes T-000098

+++ ports/rust-mt/src/main.rs
@@ -1486,7 +1486,7 @@ fn cmd_pick(
     map_set_string(&mut meta, "updated", &today_str());
     meta.insert(
         Value::String("score".to_string()),
-        Value::Number(serde_yaml::Number::from(score as i64)),
+        Value::Number(serde_yaml::Number::from(score)),
     );
 
     write_ticket(&path, &meta, &body)?;
@@ -1646,7 +1646,7 @@ fn cmd_allocate_task(
     map_set_string(&mut meta, "branch", &chosen_branch);
     meta.insert(
         Value::String("score".to_string()),
-        Value::Number(serde_yaml::Number::from(score as i64)),
+        Value::Number(serde_yaml::Number::from(score)),
     );
 
     write_ticket(&path, &meta, &body)?;
